fix: async code style and move placement control
